Move enums into an Enum namespace and rename abstract base class
