+++ src/API/Util/PaymentMethods/Enums/Wallets.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OnPay\API\Util\PaymentMethods;
+namespace OnPay\API\Util\PaymentMethods\Enums;
 
 final class Wallets {
     const APPLE_PAY = 'applepay';
