+++ src/API/Util/PaymentMethods/Methods/Anyday.php
@@ -3,12 +3,12 @@
 namespace OnPay\API\Util\PaymentMethods\Methods;
 
 use OnPay\API\Util\CurrencyCodes;
-use OnPay\API\Util\PaymentMethods\Acquirers;
+use OnPay\API\Util\PaymentMethods\Enums\Acquirers;
 
 /**
  * @internal Internal use only
  */
-final class Anyday extends AbstractMethods {
+final class Anyday extends PaymentMethodAbstract {
     const CURRENCIES = [CurrencyCodes::DKK];
     const METHOD_NAME = Acquirers::ANYDAY;
 }
