+++ src/API/Util/PaymentMethods/Methods/Card.php
@@ -3,12 +3,12 @@
 namespace OnPay\API\Util\PaymentMethods\Methods;
 
 use OnPay\API\Util\CurrencyCodes;
-use OnPay\API\Util\PaymentMethods\Methods;
+use OnPay\API\Util\PaymentMethods\Enums\Methods;
 
 /**
  * @internal Internal use only
  */
-final class Card extends AbstractMethods {
+final class Card extends PaymentMethodAbstract {
     const CURRENCIES = [CurrencyCodes::ALL_CURRENCY_CODES];
     const METHOD_NAME = Methods::CARD;
 }
