+++ src/API/Util/PaymentMethods/Methods/Googlepay.php
@@ -3,12 +3,12 @@
 namespace OnPay\API\Util\PaymentMethods\Methods;
 
 use OnPay\API\Util\CurrencyCodes;
-use OnPay\API\Util\PaymentMethods\Wallets;
+use OnPay\API\Util\PaymentMethods\Enums\Wallets;
 
 /**
  * @internal Internal use only
  */
-final class Googlepay extends AbstractMethods {
+final class Googlepay extends PaymentMethodAbstract {
     const CURRENCIES = [CurrencyCodes::ALL_CURRENCY_CODES];
     const METHOD_NAME = Wallets::GOOGLE_PAY;
 }
