+++ src/API/Util/PaymentMethods/Methods/Mobilepay.php
@@ -3,12 +3,12 @@
 namespace OnPay\API\Util\PaymentMethods\Methods;
 
 use OnPay\API\Util\CurrencyCodes;
-use OnPay\API\Util\PaymentMethods\Wallets;
+use OnPay\API\Util\PaymentMethods\Enums\Wallets;
 
 /**
  * @internal Internal use only
  */
-final class Mobilepay extends AbstractMethods {
+final class Mobilepay extends PaymentMethodAbstract {
     const CURRENCIES = [
         CurrencyCodes::DKK,
         CurrencyCodes::NOK,
