+++ src/API/Util/PaymentMethods/Methods/Paypal.php
@@ -3,12 +3,12 @@
 namespace OnPay\API\Util\PaymentMethods\Methods;
 
 use OnPay\API\Util\CurrencyCodes;
-use OnPay\API\Util\PaymentMethods\Acquirers;
+use OnPay\API\Util\PaymentMethods\Enums\Acquirers;
 
 /**
  * @internal Internal use only
  */
-final class Paypal extends AbstractMethods {
+final class Paypal extends PaymentMethodAbstract {
     const CURRENCIES = [CurrencyCodes::ALL_CURRENCY_CODES];
     const METHOD_NAME = Acquirers::PAYPAL;
 }
