+++ src/API/Util/PaymentMethods/Methods/Swish.php
@@ -3,12 +3,12 @@
 namespace OnPay\API\Util\PaymentMethods\Methods;
 
 use OnPay\API\Util\CurrencyCodes;
-use OnPay\API\Util\PaymentMethods\Acquirers;
+use OnPay\API\Util\PaymentMethods\Enums\Acquirers;
 
 /**
  * @internal Internal use only
  */
-final class Swish extends AbstractMethods {
+final class Swish extends PaymentMethodAbstract {
     const CURRENCIES = [CurrencyCodes::SEK];
     const METHOD_NAME = Acquirers::SWISH;
 }
