+++ src/API/Util/PaymentMethods/Methods/Vipps.php
@@ -3,12 +3,12 @@
 namespace OnPay\API\Util\PaymentMethods\Methods;
 
 use OnPay\API\Util\CurrencyCodes;
-use OnPay\API\Util\PaymentMethods\Wallets;
+use OnPay\API\Util\PaymentMethods\Enums\Wallets;
 
 /**
  * @internal Internal use only
  */
-final class Vipps extends AbstractMethods {
+final class Vipps extends PaymentMethodAbstract {
     const CURRENCIES = [CurrencyCodes::NOK];
     const METHOD_NAME = Wallets::VIPPS;
 }
